♻️ turn `qfr_dd` into a real library with compiled sources
